+++ auto/restore/config_test.go
@@ -1,4 +1,4 @@
-package download
+package restore
 
 import (
 	"bytes"
@@ -9,6 +9,7 @@ import (
 	"testing"
 	"time"
 
+	"github.com/rqlite/rqlite/auto"
 	"github.com/rqlite/rqlite/aws"
 )
 
@@ -135,7 +136,7 @@ func TestUnmarshal(t *testing.T) {
 			expectedCfg: &Config{
 				Version:           1,
 				Type:              "s3",
-				Timeout:           30 * Duration(time.Second),
+				Timeout:           30 * auto.Duration(time.Second),
 				ContinueOnFailure: false,
 			},
 			expectedS3: &aws.S3Config{
@@ -166,7 +167,7 @@ func TestUnmarshal(t *testing.T) {
 			expectedCfg: &Config{
 				Version:           1,
 				Type:              "s3",
-				Timeout:           Duration(30 * time.Second),
+				Timeout:           auto.Duration(30 * time.Second),
 				ContinueOnFailure: true,
 			},
 			expectedS3: &aws.S3Config{
@@ -195,7 +196,7 @@ func TestUnmarshal(t *testing.T) {
 			}			`),
 			expectedCfg: nil,
 			expectedS3:  nil,
-			expectedErr: ErrInvalidVersion,
+			expectedErr: auto.ErrInvalidVersion,
 		},
 		{
 			name: "UnsupportedType",
@@ -214,7 +215,7 @@ func TestUnmarshal(t *testing.T) {
 			}			`),
 			expectedCfg: nil,
 			expectedS3:  nil,
-			expectedErr: ErrUnsupportedStorageType,
+			expectedErr: auto.ErrUnsupportedStorageType,
 		},
 	}