+++ auto/types.go
@@ -1,18 +1,14 @@
-package upload
+package auto
 
 import (
 	"encoding/json"
 	"errors"
-	"io/ioutil"
-	"os"
 	"time"
-
-	"github.com/rqlite/rqlite/aws"
 )
 
 const (
-	// version is the max version of the config file format supported
-	version = 1
+	// Version is the max version of the config file format supported
+	Version = 1
 )
 
 var (
@@ -72,51 +68,4 @@ func (s *StorageType) UnmarshalJSON(b []byte) error {
 	default:
 		return ErrUnsupportedStorageType
 	}
-}
-
-// Config is the config file format for the upload service
-type Config struct {
-	Version    int             `json:"version"`
-	Type       StorageType     `json:"type"`
-	NoCompress bool            `json:"no_compress,omitempty"`
-	Interval   Duration        `json:"interval"`
-	Sub        json.RawMessage `json:"sub"`
-}
-
-// Unmarshal unmarshals the config file and returns the config and subconfig
-func Unmarshal(data []byte) (*Config, *aws.S3Config, error) {
-	cfg := &Config{}
-	err := json.Unmarshal(data, cfg)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	if cfg.Version > version {
-		return nil, nil, ErrInvalidVersion
-	}
-
-	s3cfg := &aws.S3Config{}
-	err = json.Unmarshal(cfg.Sub, s3cfg)
-	if err != nil {
-		return nil, nil, err
-	}
-	return cfg, s3cfg, nil
-}
-
-// ReadConfigFile reads the config file and returns the data. It also expands
-// any environment variables in the config file.
-func ReadConfigFile(filename string) ([]byte, error) {
-	f, err := os.Open(filename)
-	if err != nil {
-		return nil, err
-	}
-	defer f.Close()
-
-	data, err := ioutil.ReadAll(f)
-	if err != nil {
-		return nil, err
-	}
-
-	data = []byte(os.ExpandEnv(string(data)))
-	return data, nil
 }